+++ deprecated/templates/githooks/README.md
