+++ deprecated/templates/githooks/post-checkout/clean-python-cache
