+++ deprecated/templates/githooks/pre-push/check-clang-format
