+++ deprecated/templates/githooks/pre-push/check-code-freeze
