+++ deprecated/templates/githooks/pre-push/check-compile
