+++ deprecated/templates/githooks/pre-push/check-for-tickets
