+++ deprecated/templates/githooks/pre-push/check-uncommitted
