+++ deprecated/templates/mongo-cmakelists.txt
